Switched to drawing to a texture and stretching it every frame.

+++ SolasLib/include/SolasLib/core/Light.hpp
@@ -13,7 +13,7 @@ public:
 	glm::vec2 direction;
 	float span;
 	bool shouldUpdate;
-	float* lightMap;
+	int* lightMap;
 	int lightMapWidth;
 	int lightMapHeight;
 	Light(float x, float y, glm::vec2 direction, float span, float range, int tileSize)
@@ -29,6 +29,6 @@ public:
 		this->lightMapWidth = TILE_BUFFER * 2 + radiusInTiles * 2;
 		this->lightMapHeight = TILE_BUFFER * 2 + radiusInTiles * 2;
 
-		this->lightMap = new float[lightMapWidth * lightMapHeight]();
+		this->lightMap = new int[lightMapWidth * lightMapHeight]();
 	}
 };
